fix(demo): carry tags, test meta and ai usage on simulated runs
